Pull the warning stuff to a new module
